v4.6.2-beta: unified parser, BIR verifier overhaul, frontend restructure, GPU frontend removal

+++ zig/src/compiler/backend/targets/x64/isel/context.zig
@@ -11,7 +11,7 @@ pub const OffsetMap = std.AutoHashMap(u32, i32);
 
 pub const BlockFixup = struct {
     disp_pos: usize,
-    target: usize,
+    target: u32,
 };
 
 pub const CallFixup = struct {
